SERCOM OK

+++ Manejador.cpp
@@ -5,20 +5,20 @@
 #include <string>
 #include <algorithm>
 #include <iostream>
+#include <unordered_map>
 #define BUENO 0
 #define CICLO -1
 #define AISLADO -2
 
 
-Manejador::Manejador (Monitor& monitor_inic, Monitor& monitor_fin)
+Manejador::Manejador(Monitor& monitor_inic, Monitor& monitor_fin)
     :   monitor_inic(monitor_inic),
         monitor_fin(monitor_fin)
 {}
-
 void Manejador::llenarGrafo(Grafo& grafo,
                             std::vector<std::vector<std::string>>& lineas,
                             std::unordered_map<std::string, int>& dic_et){
-    for (int i = 0; i < lineas.size(); i++) {
+    for (unsigned int i = 0; i < lineas.size(); i++) {
         std::string instr_act = lineas[i][0];
         if (instr_act.find(':') == std::string::npos) { // si no tiene etiqueta
             if (lineas[i][0].compare("ret") == 0 || (i + 1) > lineas.size() - 1)
@@ -27,12 +27,12 @@ void Manejador::llenarGrafo(Grafo& grafo,
         } else{
             if (lineas[i][1].compare("ret") == 0 || (i + 1) > lineas.size() - 1)
                 continue;
-            revisarOpcode( i, 1, lineas, dic_et, grafo);
+            revisarOpcode(i, 1, lineas, dic_et, grafo);
         }
     }
 }
 
-void Manejador::revisarOpcode(int i, int pos_instr,
+void Manejador::revisarOpcode(size_t i, size_t pos_instr,
                               std::vector<std::vector<std::string>>& lineas,
                               std::unordered_map<std::string, int>& dic_et,
                               Grafo& grafo){
@@ -47,9 +47,10 @@ void Manejador::revisarOpcode(int i, int pos_instr,
         } else if ((lineas[i]).size() == pos_instr + 4) {
             std::string etiqueta_a_parsear = lineas[i][pos_instr + 2];
             std::string etiqueta_act2 = lineas[i][pos_instr + 3];
-            etiqueta_act = etiqueta_a_parsear.substr(0,etiqueta_a_parsear.size()-1);
+            etiqueta_act = etiqueta_a_parsear.substr(0,
+                                             etiqueta_a_parsear.size() - 1);
             int aux = buscarDic(dic_et, etiqueta_act2);
-            if(aux != -1) grafo.agregarArista(i, aux);
+            if (aux != -1) grafo.agregarArista(i, aux);
         }
         int aux = buscarDic(dic_et, etiqueta_act);
         if (aux != -1)
@@ -68,7 +69,6 @@ int Manejador::buscarDic(std::unordered_map<std::string, int>& dic_et,
 }
 
 void Manejador::run(){
-
     while (!monitor_inic.vacia()) {
         std::vector<std::vector<std::string>> lineas;
         std::unordered_map<std::string, int> dic_et;
@@ -80,7 +80,6 @@ void Manejador::run(){
         llenarGrafo(grafo, lineas, dic_et);
         mandarMensaje(grafo, archivo_actual);
     }
-
 }
 
 void Manejador::mandarMensaje(Grafo& grafo, std::string archivo_actual){
